Extract the parser
