+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "lsp/nls",
     "nickel",
     "package",
+    "parser",
     "py-nickel",
     "utils",
     "vector",
@@ -31,6 +32,7 @@ nickel-lang-core = { version = "0.15.0", path = "./core", default-features = fal
 nickel-lang-flock = { version = "0.1.0", path = "./flock" }
 nickel-lang-git = { version = "0.1.0", path = "./git" }
 nickel-lang-package = { version = "0.4.0", path = "./package" }
+nickel-lang-parser = { version = "0.1.0", path = "./parser" }
 nickel-lang-vector = { version = "0.1.0", path = "./vector" }
 nickel-lang-utils = { version = "0.1.0", path = "./utils" }
 lsp-harness = { version = "0.1.0", path = "./lsp/lsp-harness" }
