+++ core/src/bytecode/mod.rs
@@ -3,5 +3,7 @@
 //! This implementation is under construction; it's currently not usable and it's not available by
 //! default in mainline Nickel.
 
+// TODO(parser migration): compatibility shim
+pub use nickel_lang_parser::pretty;
+
 pub mod ast;
-pub mod pretty;
