+++ core/src/label.rs
@@ -18,6 +18,9 @@ use crate::{
     typ::{Type, TypeF},
 };
 
+// TODO(parser migration): compatibility shim
+pub use nickel_lang_parser::ast::MergeKind;
+
 pub mod ty_path {
     //! Type paths.
     //!
@@ -527,21 +530,6 @@ impl Default for Label {
             field_name: None,
         }
     }
-}
-
-/// Possible origins of a merge operation.
-#[derive(Clone, Copy, Eq, PartialEq, Debug, Default)]
-pub enum MergeKind {
-    /// A standard, user-written merge operation (or a merge operation descending from a
-    /// user-written merge operation).
-    #[default]
-    Standard,
-    /// A merge generated by the parser when reconstructing piecewise definition, for example:
-    ///
-    /// ```nickel
-    /// { foo = def1, foo = def2}
-    /// ```
-    PiecewiseDef,
 }
 
 /// A merge label.
