+++ core/src/lib.rs
@@ -3,16 +3,17 @@
 // but if not then we should reduce them.
 #![allow(clippy::result_large_err)]
 
+// TODO(parser migration): compatibility shim
+use nickel_lang_parser::metrics;
+pub use nickel_lang_parser::{environment, files, identifier, traverse};
+
 pub mod bytecode;
 pub mod cache;
 pub mod closurize;
 pub mod combine;
 pub mod deserialize;
-pub mod environment;
 pub mod error;
 pub mod eval;
-pub mod files;
-pub mod identifier;
 pub mod label;
 #[cfg(feature = "nix-experimental")]
 pub mod nix_ffi;
@@ -26,11 +27,8 @@ pub mod serialize;
 pub mod stdlib;
 pub mod term;
 pub mod transform;
-pub mod traverse;
 pub mod typ;
 pub mod typecheck;
-
-pub(crate) mod metrics;
 
 #[cfg(feature = "format")]
 pub mod format;
