+++ parser/src/lib.rs
@@ -1,15 +1,8 @@
 use crate::{
-    bytecode::ast::{
-        Ast, AstAlloc,
-        compat::{FromAst, ToMainline},
-        typ::Type,
-    },
+    ast::{Ast, AstAlloc, typ::Type},
     error::{ParseError, ParseErrors},
-    eval::value::NickelValue,
     files::FileId,
     identifier::LocIdent,
-    metrics,
-    position::PosTable,
     position::RawSpan,
 };
 
@@ -19,17 +12,43 @@ lalrpop_mod!(
     #[allow(clippy::all)]
     #[allow(unused_parens)]
     #[allow(unused_imports)]
-    pub grammar, "/parser/grammar.rs");
+    pub grammar, "/grammar.rs");
 
 use grammar::__ToTriple;
 
+pub mod ast;
+pub mod combine;
+pub mod environment;
 pub mod error;
+pub mod files;
+pub mod identifier;
 pub mod lexer;
+pub mod metrics;
+pub mod position;
+pub mod pretty;
+pub mod traverse;
+pub mod typ;
 pub(crate) mod uniterm;
 pub mod utils;
 
 #[cfg(test)]
 mod tests;
+
+/// The underlying type representing Nickel numbers. Currently, numbers are arbitrary precision
+/// rationals.
+///
+/// Basic arithmetic operations are exact, without loss of precision (within the limits of available
+/// memory).
+///
+/// Raising to a power that doesn't fit in a signed 64bits number will lead to converting both
+/// operands to 64-bits floats, performing the floating-point power operation, and converting back
+/// to rationals, which can incur a loss of precision.
+///
+/// [^number-serialization]: Conversion to string and serialization try to first convert the
+///     rational as an exact signed or usigned 64-bits integer. If this succeeds, such operations
+///     don't lose precision. Otherwise, the number is converted to the nearest 64bit float and then
+///     serialized/printed, which can incur a loss of information.
+pub type Number = malachite::rational::Rational;
 
 /// Either an expression or a toplevel let declaration.
 ///
@@ -60,11 +79,14 @@ trait LalrpopParser<'ast, T> {
         alloc: &'ast AstAlloc,
         src_id: FileId,
         errors: &'err mut Vec<
-            lalrpop_util::ErrorRecovery<usize, lexer::Token<'input>, self::error::ParseError>,
+            lalrpop_util::ErrorRecovery<usize, lexer::Token<'input>, self::error::ParseOrLexError>,
         >,
         next_wildcard_id: &'wcard mut usize,
         __tokens0: __TOKENS,
-    ) -> Result<T, lalrpop_util::ParseError<usize, lexer::Token<'input>, self::error::ParseError>>
+    ) -> Result<
+        T,
+        lalrpop_util::ParseError<usize, lexer::Token<'input>, self::error::ParseOrLexError>,
+    >
     where
         __TOKEN: __ToTriple<'input, 'ast, 'err, 'wcard>,
         __TOKENS: IntoIterator<Item = __TOKEN>;
@@ -83,14 +105,14 @@ macro_rules! generate_lalrpop_parser_impl {
                     lalrpop_util::ErrorRecovery<
                         usize,
                         lexer::Token<'input>,
-                        self::error::ParseError,
+                        self::error::ParseOrLexError,
                     >,
                 >,
                 next_wildcard_id: &'wcard mut usize,
                 __tokens0: __TOKENS,
             ) -> Result<
                 $output,
-                lalrpop_util::ParseError<usize, lexer::Token<'input>, self::error::ParseError>,
+                lalrpop_util::ParseError<usize, lexer::Token<'input>, self::error::ParseOrLexError>,
             >
             where
                 __TOKEN: __ToTriple<'input, 'ast, 'err, 'wcard>,
@@ -121,7 +143,7 @@ pub trait ErrorTolerantParser<'ast, T> {
         &self,
         alloc: &'ast AstAlloc,
         file_id: FileId,
-        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::ParseError>>,
+        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::LexicalError>>,
     ) -> Result<(T, ParseErrors), ParseError>;
 
     /// Parse a value from a lexer with the given `file_id`, failing at the first encountered
@@ -130,7 +152,7 @@ pub trait ErrorTolerantParser<'ast, T> {
         &self,
         alloc: &'ast AstAlloc,
         file_id: FileId,
-        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::ParseError>>,
+        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::LexicalError>>,
     ) -> Result<T, ParseErrors>;
 }
 
@@ -142,7 +164,7 @@ where
         &self,
         alloc: &'ast AstAlloc,
         file_id: FileId,
-        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::ParseError>>,
+        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::LexicalError>>,
     ) -> Result<(T, ParseErrors), ParseError> {
         let mut parse_errors = Vec::new();
         let mut next_wildcard_id = 0;
@@ -152,7 +174,7 @@ where
                 file_id,
                 &mut parse_errors,
                 &mut next_wildcard_id,
-                lexer,
+                lexer.map(|x| x.map_err(error::ParseOrLexError::from)),
             )
             .map_err(|err| ParseError::from_lalrpop(err, file_id));
 
@@ -167,7 +189,7 @@ where
         &self,
         alloc: &'ast AstAlloc,
         file_id: FileId,
-        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::ParseError>>,
+        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::LexicalError>>,
     ) -> Result<T, ParseErrors> {
         match self.parse_tolerant(alloc, file_id, lexer) {
             Ok((t, e)) if e.no_errors() => Ok(t),
@@ -199,7 +221,7 @@ pub trait FullyErrorTolerantParser<'ast, T> {
         &self,
         alloc: &'ast AstAlloc,
         file_id: FileId,
-        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::ParseError>>,
+        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::LexicalError>>,
         full_span: RawSpan,
     ) -> (T, ParseErrors);
 }
@@ -212,7 +234,7 @@ where
         &self,
         alloc: &'ast AstAlloc,
         file_id: FileId,
-        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::ParseError>>,
+        lexer: impl Iterator<Item = Result<lexer::SpannedToken<'input>, error::LexicalError>>,
         full_span: RawSpan,
     ) -> (Ast<'ast>, ParseErrors) {
         match self.parse_tolerant(alloc, file_id, lexer) {
@@ -222,129 +244,5 @@ where
                 (ast, e.into())
             }
         }
-    }
-}
-
-/// General interface of the various specialized Nickel parsers.
-///
-/// This trait is a compatibility layer version of [ErrorTolerantParser]. It produces data of the
-/// old, mainline types because the current pipeline still depends on them (defined in
-/// [crate::term]). Eventually we'll get rid of it and only use [ErrorTolerantParser], which
-/// produces the new AST instead.
-pub trait ErrorTolerantParserCompat<T> {
-    /// Parse a value from a lexer with the given `file_id` in an error-tolerant way. This methods
-    /// can still fail for non-recoverable errors.
-    fn parse_tolerant_compat(
-        &self,
-        pos_table: &mut PosTable,
-        file_id: FileId,
-        lexer: lexer::Lexer,
-    ) -> Result<(T, ParseErrors), ParseError>;
-
-    /// Parse a value from a lexer with the given `file_id`, failing at the first encountered
-    /// error.
-    fn parse_strict_compat(
-        &self,
-        pos_table: &mut PosTable,
-        file_id: FileId,
-        lexer: lexer::Lexer,
-    ) -> Result<T, ParseErrors>;
-}
-
-impl<'ast> FromAst<ExtendedTerm<Ast<'ast>>> for ExtendedTerm<NickelValue> {
-    fn from_ast(ast: &ExtendedTerm<Ast<'ast>>, pos_table: &mut PosTable) -> Self {
-        match ast {
-            ExtendedTerm::Term(t) => ExtendedTerm::Term(t.to_mainline(pos_table)),
-            ExtendedTerm::ToplevelLet(ident, t) => {
-                ExtendedTerm::ToplevelLet(*ident, t.to_mainline(pos_table))
-            }
-        }
-    }
-}
-
-// Generate boilerplate impl to produce legacy mainline types from the available parsers.
-macro_rules! generate_compat_impl {
-    ($parser:ty, $output:ty) => {
-        impl ErrorTolerantParserCompat<$output> for $parser {
-            fn parse_tolerant_compat(
-                &self,
-                pos_table: &mut PosTable,
-                file_id: FileId,
-                lexer: lexer::Lexer,
-            ) -> Result<($output, ParseErrors), ParseError> {
-                let alloc = AstAlloc::new();
-                self.parse_tolerant(&alloc, file_id, lexer).map(|(t, e)| {
-                    (
-                        metrics::measure_runtime!(
-                            "runtime:ast_conversion",
-                            t.to_mainline(pos_table)
-                        ),
-                        e,
-                    )
-                })
-            }
-
-            fn parse_strict_compat(
-                &self,
-                pos_table: &mut PosTable,
-                file_id: FileId,
-                lexer: lexer::Lexer,
-            ) -> Result<$output, ParseErrors> {
-                let alloc = AstAlloc::new();
-                self.parse_strict(&alloc, file_id, lexer).map(|t| {
-                    metrics::measure_runtime!("runtime:ast_conversion", t.to_mainline(pos_table))
-                })
-            }
-        }
-    };
-}
-
-generate_compat_impl!(grammar::ExtendedTermParser, ExtendedTerm<NickelValue>);
-generate_compat_impl!(grammar::TermParser, NickelValue);
-generate_compat_impl!(grammar::FixedTypeParser, crate::typ::Type);
-
-impl ErrorTolerantParserCompat<(Vec<LocIdent>, NickelValue, RawSpan)>
-    for grammar::CliFieldAssignmentParser
-{
-    fn parse_tolerant_compat(
-        &self,
-        pos_table: &mut PosTable,
-        file_id: FileId,
-        lexer: lexer::Lexer,
-    ) -> Result<((Vec<LocIdent>, NickelValue, RawSpan), ParseErrors), ParseError> {
-        self.parse_tolerant(&AstAlloc::new(), file_id, lexer)
-            .map(|((path, term, span), e)| ((path, term.to_mainline(pos_table), span), e))
-    }
-
-    fn parse_strict_compat(
-        &self,
-        pos_table: &mut PosTable,
-        file_id: FileId,
-        lexer: lexer::Lexer,
-    ) -> Result<(Vec<LocIdent>, NickelValue, RawSpan), ParseErrors> {
-        self.parse_strict(&AstAlloc::new(), file_id, lexer)
-            .map(|(path, term, span)| (path, term.to_mainline(pos_table), span))
-    }
-}
-
-// This implementation doesn't do any conversion, but hide away the (useless, in this case)
-// [crate::bytecode::ast::AstAlloc] parameter.
-impl ErrorTolerantParserCompat<Vec<LocIdent>> for grammar::StaticFieldPathParser {
-    fn parse_tolerant_compat(
-        &self,
-        _pos_table: &mut PosTable,
-        file_id: FileId,
-        lexer: lexer::Lexer,
-    ) -> Result<(Vec<LocIdent>, ParseErrors), ParseError> {
-        self.parse_tolerant(&AstAlloc::new(), file_id, lexer)
-    }
-
-    fn parse_strict_compat(
-        &self,
-        _pos_table: &mut PosTable,
-        file_id: FileId,
-        lexer: lexer::Lexer,
-    ) -> Result<Vec<LocIdent>, ParseErrors> {
-        self.parse_strict(&AstAlloc::new(), file_id, lexer)
     }
 }
